+++ packages/rx_bloc_cli/mason_templates/bricks/lib_social_logins/__brick__/lib/lib_social_logins/repositories/google_auth_repository.dart
@@ -1,0 +1,32 @@
+{{> licence.dart }}
+
+import '../../base/common_mappers/error_mappers/error_mapper.dart';
+import '../../lib_auth/models/auth_token_model.dart';
+import '../data_sources/google_auth_data_source.dart';
+import '../data_sources/google_credential_data_source.dart';
+import '../models/google_auth_request_model.dart';
+import '../models/google_credentials_model.dart';
+
+class GoogleAuthRepository {
+  GoogleAuthRepository(this._googleAuthDataSource, this._errorMapper,
+      this._googleCredentialDataSource);
+  final GoogleAuthDataSource _googleAuthDataSource;
+  final ErrorMapper _errorMapper;
+  final GoogleCredentialDataSource _googleCredentialDataSource;
+
+  Future<AuthTokenModel> googleAuth({
+    required GoogleCredentialsModel googleAuthRequestModel,
+  }) =>
+      _errorMapper.execute(
+        () => _googleAuthDataSource.googleAuth(
+            GoogleAuthRequestModel.fromGoogleCredentials(
+                googleAuthRequestModel)),
+      );
+
+  Future<GoogleCredentialsModel> getUsersGoogleCredential() =>
+      _errorMapper.execute(() async {
+        final credentials =
+            await _googleCredentialDataSource.getUsersGoogleCredential();
+        return GoogleCredentialsModel.fromGoogleCredentials(credentials!);
+      });
+}