+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/app/{{project_name}}.dart
@@ -1,14 +1,17 @@
 {{> licence.dart }}
+{{#enable_change_language}}
+import 'dart:async'; {{/enable_change_language}}
 {{#analytics}}
 import 'package:firebase_analytics/firebase_analytics.dart';{{/analytics}}{{#push_notifications}}
 import 'package:firebase_messaging/firebase_messaging.dart';{{/push_notifications}}
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:provider/provider.dart';
-
+import 'package:provider/provider.dart'; {{#enable_change_language}}
+import 'package:widget_toolkit/language_picker.dart'; {{/enable_change_language}}
 import '../../l10n/l10n.dart';
-import '../../lib_auth/data_sources/remote/interceptors/auth_interceptor.dart';
+import '../../lib_auth/data_sources/remote/interceptors/auth_interceptor.dart'; {{#enable_change_language}}
+import '../../lib_change_language/bloc/change_language_bloc.dart';{{/enable_change_language}}
 import '../../lib_router/router.dart';
 import '../data_sources/remote/http_clients/api_http_client.dart';
 import '../data_sources/remote/http_clients/plain_http_client.dart';{{#analytics}}
@@ -43,20 +46,47 @@ class {{project_name.pascalCase()}} extends StatelessWidget {
 class _MyMaterialApp extends StatefulWidget {
   const _MyMaterialApp();
 
-@override
-__MyMaterialAppState createState() => __MyMaterialAppState();
+  @override
+  __MyMaterialAppState createState() => __MyMaterialAppState();
 }
 
 class __MyMaterialAppState extends State<_MyMaterialApp> {
+  Locale? _locale;
+  {{#enable_change_language}}
+  late StreamSubscription<LanguageModel> _languageSubscription; {{/enable_change_language}}
 
   @override
-  void initState() { {{#push_notifications}}
+  void initState() {
+    {{^enable_change_language}}
+    _locale = const Locale('en'); {{/enable_change_language}}
+    {{#enable_change_language}}
+    _updateLocale(); {{/enable_change_language}} {{#push_notifications}}
     _configureFCM(); {{/push_notifications}}
     _configureInterceptors();
-
     super.initState();
-  }{{#push_notifications}}
+  }
 
+  {{#enable_change_language}}
+  void _updateLocale() {
+    _languageSubscription = context
+        .read<ChangeLanguageBlocType>()
+        .states
+        .currentLanguage
+        .listen((language) {
+      setState(
+        () => _locale = Locale(language.locale),
+      );
+    });
+  } {{/enable_change_language}}
+
+  {{#enable_change_language}}
+  @override
+  void dispose() {
+    _languageSubscription.cancel();
+    super.dispose();
+  } {{/enable_change_language}}
+
+  {{#push_notifications}}
   Future<void> _configureFCM() async {
     /// Initialize the FCM callbacks
     if (kIsWeb){
@@ -98,9 +128,10 @@ class __MyMaterialAppState extends State<_MyMaterialApp> {
         darkTheme: {{project_name.pascalCase()}}Theme.buildTheme(DesignSystem.dark()),
         localizationsDelegates: const [
           I18n.delegate,
-          GlobalMaterialLocalizations.delegate,
+          ...GlobalMaterialLocalizations.delegates,
         ],
         supportedLocales: I18n.supportedLocales,
+        locale: _locale,
         routerConfig: context.read<AppRouter>().router,
         debugShowCheckedModeBanner: false,
       );