+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/theme/design_system/design_system_typography.dart
@@ -90,4 +90,21 @@ class DesignSystemTypography {
 {{/enable_feature_counter}}
   late final fadedButtonText =
       h3Med14.copyWith(color: _designSystemColor.black);
+  {{#enable_social_logins}}
+  late final googleButtonText = TextStyle(
+    fontSize: 14,
+    backgroundColor: _designSystemColor.googleBackground,
+    color: _designSystemColor.googleButtonText,
+  );
+
+  late final facebookButtonText = TextStyle(
+    fontSize: 14,
+    backgroundColor: _designSystemColor.facebookBackground,
+    color: _designSystemColor.facebookTextColor,
+  );
+  late final appleButtonText = TextStyle(
+    fontSize: 14,
+    color: _designSystemColor.appleButtonText,
+  );
+  {{/enable_social_logins}}
 }