+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/feature_login/views/login_page.dart
@@ -4,7 +4,9 @@ import 'package:flutter/material.dart';
 
 import '../../app_extensions.dart';
 import '../../base/common_ui_components/custom_app_bar.dart';{{#enable_social_logins}}
-import '../../lib_social_logins/ui_components/apple_login_widget.dart';{{/enable_social_logins}}
+import '../../lib_social_logins/ui_components/apple_login_widget.dart';
+import '../../lib_social_logins/ui_components/facebook_login_widget.dart';
+import '../../lib_social_logins/ui_components/google_login_widget.dart';{{/enable_social_logins}}
 import '../ui_components/login_form.dart';
 
 class LoginPage extends StatelessWidget {
@@ -17,33 +19,25 @@ class LoginPage extends StatelessWidget {
           title: context.l10n.featureLogin.loginPageTitle,
         ),
         body: SafeArea(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Center(
-                child: Container(
-                  width: context.designSystem.spacing.xxxxl300,
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(
-                      context.designSystem.spacing.xsss,
-                    ),
-                    border: Border.all(
-                      color: Theme.of(context).colorScheme.onBackground,
-                    ),
-                  ),
-                  child: LoginForm(
-                    title: context.l10n.featureLogin.loginCredentialsHint,
-                  ),
-                ),
-              ),{{#enable_social_logins}}
-              Padding(
-                padding: EdgeInsets.symmetric(
-                    horizontal: context.designSystem.spacing.xxl,
-                    vertical: context.designSystem.spacing.l,
-                ),
-                child: const AppleLoginWidget(),
-              ),{{/enable_social_logins}}
-            ],
+          child: Padding(
+            padding: EdgeInsets.symmetric(
+              horizontal: context.designSystem.spacing.xxl2,
+            ),
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: [
+                LoginForm(
+                  title: context.l10n.featureLogin.loginCredentialsHint,
+                ),{{#enable_social_logins}}
+                SizedBox(height: context.designSystem.spacing.xs),
+                const FacebookLoginWidget(),
+                SizedBox(height: context.designSystem.spacing.xs),
+                const AppleLoginWidget(),
+                SizedBox(height: context.designSystem.spacing.xs),
+                const GoogleLoginWidget(),{{/enable_social_logins}}
+              ],
+            ),
           ),
         ),
       );